refactor GudaLifeBoard to split out the state mgmt into new GudaLifeDict
Refactored GudaLifeBoard so that a new GudaLifeDict could be used as the state management.  This will allow using a different backend to manage the state, such as using a pure python list of lists as a sparse matrix representation.  Will allow using NumPy arrays or for direct memory mgmt.
Refactored the test cases to reflect new refactoring.